Build the PhysicsServer3DExtension against godot-cpp 4.2
Match the generated virtual signatures (_step takes float, the soft-body point
index is int32_t) and mint RIDs with UtilityFunctions::rid_from_int64 rather than
a nonexistent RID::from_uint64. The bridge stays 0-based as proved; the engine
RID is offset by one so id 0 remains invalid, and _free_rid recovers the 0-based
id on the way back. The extension now links to a loadable DLL, with the
gdextension config that points Godot at it.

+++ src/mujoco_physics_server.h
@@ -41,7 +41,7 @@ public:
 	virtual void _init() override;
 	virtual void _finish() override;
 	virtual void _set_active(bool p_active) override;
-	virtual void _step(double p_step) override;
+	virtual void _step(float p_step) override;
 	virtual void _sync() override;
 	virtual void _end_sync() override;
 	virtual void _flush_queries() override;
@@ -63,7 +63,7 @@ public:
 	// --- soft bodies (cloth -> MuJoCo flex) ---
 	virtual RID _soft_body_create() override;
 	virtual Vector3 _soft_body_get_point_global_position(const RID &p_body,
-			int64_t p_point_index) const override;
+			int32_t p_point_index) const override;
 
 	// --- teardown ---
 	virtual void _free_rid(const RID &p_rid) override;
